requirements/base.txt
argon2-cffi==21.3.0
asgiref==3.4.1
dj-database-url==0.5.0
dj-rest-auth==2.2.2
Django==3.2.11
django-allauth==0.47.0
django-cors-headers==3.11.0
django-filter==21.1
django-model-utils==4.2.0
# django-polymorphic==3.1.0
# django-redis==5.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
Pillow==9.0.0
psycopg2-binary==2.9.3
python-decouple==3.5
pytz==2021.3
